SVG conversion
Convert SVG to image.
The <svg> tag can't be "right-click, copied" to be used in a report or else.
This change encode the svg in base64 in an image and can be copied/saved as for other uses.

It also embed the css style, but with a pre-defined set of attributes. This will need adjusting if other attributes are used.